Changes made in this commit: - Renamed `CliConfig.kt` to `CliConfigManager.kt` with a similarity index of 100%. - Deleted `CommitConfig.kt` file. - Added new file `SummaryOptions.kt` under `core` package. - Renamed `GitOperations.kt` to `GitChangesSummarizer.kt` with a similarity index of 67%. - Renamed `GitService.kt` to `GitService.kt`. - Renamed `ProgressBarGui.kt` to `ProgressBarGui.kt`. - Renamed `TerminalGui.kt` to `TerminalGui.kt` with a similarity index of 97%. - Renamed `AnthropicLlmResponse.kt` to `AnthropicLlmProvider.kt` with a similarity index of 58%. - Added new file `GoogleVertexLlmProvider.kt` under `llm.provider` package. - Added new file `LocalLlmProvider.kt` under `llm.provider` package. - Renamed `OpenaiLlmResponse.kt` to `OpenaiLlmProvider.kt` with a similarity index of 60%. - Moved `AnthropicLlmResponseService.kt` to `llm/service` package with a similarity index of 56%. - Moved `DefaultLlmResponseService.kt` to `llm/service` package with a similarity index of 56%. - Added new file `LlmResponseService.kt` under `services` package. - Deleted `GoogleVertexLlmResponse.kt` file. - Deleted `LlmConstants.kt` file. - Deleted `LlmResponseService.kt` file. - Deleted `LocalLlmResponse.kt` file.
