--- src/main/kotlin/com/ibrahimharoon/gitautocommit/cli/CommitConfig.kt
@@ -1,10 +1,0 @@
-package com.ibrahimharoon.gitautocommit.cli
-
-data class CommitConfig(
-    val useLocal: Boolean,
-    val useOpenai: Boolean,
-    val useAnthropic: Boolean,
-    val useGoogle: Boolean,
-    val isPr: Boolean,
-    val isPlainPr: Boolean
-)
