+++ src/main/kotlin/com/ibrahimharoon/gitautocommit/core/SummaryOptions.kt
@@ -1,0 +1,9 @@
+package com.ibrahimharoon.gitautocommit.cli
+
+import com.ibrahimharoon.gitautocommit.llm.provider.LlmProvider
+
+data class SummaryOptions(
+    val llmProvider: LlmProvider,
+    val isPr: Boolean,
+    val isPlainPr: Boolean
+)
