+++ src/main/kotlin/com/ibrahimharoon/gitautocommit/llm/service/AnthropicLlmResponseService.kt
@@ -1,10 +1,6 @@
 package com.ibrahimharoon.gitautocommit.services
 
-import com.ibrahimharoon.gitautocommit.cache.RegenConversationCache
 import com.ibrahimharoon.gitautocommit.rest.dtos.AnthropicLlmResponseDto
-import com.ibrahimharoon.gitautocommit.templates.LlmTemplates.Companion.COMMIT_PROMPT
-import com.ibrahimharoon.gitautocommit.templates.LlmTemplates.Companion.PR_SUMMARY_PROMPT
-import com.ibrahimharoon.gitautocommit.templates.LlmTemplates.Companion.ROLE
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpMethod
@@ -16,27 +12,18 @@ class AnthropicLlmResponseService(
     private val url: String,
     private val headers: HttpHeaders
 ) : LlmResponseService {
-    private val conversationCache = RegenConversationCache
-
-    override fun getMessage(gitData: String, isPr: Boolean, additionalLlmPrompt: String): String {
-        var prompt = ""
-        if (additionalLlmPrompt.isNotEmpty()) {
-            prompt = "Your previous commit message was no good. This is an additional prompt to help you out: $additionalLlmPrompt\n\n"
-        }
-        prompt += if (isPr) PR_SUMMARY_PROMPT else COMMIT_PROMPT
-        prompt += if (conversationCache.isNotEmpty()) "Previous history: $conversationCache" else ""
-
+    override fun response(role: String, prompt: String): String {
         val messages = listOf(
             mapOf(
                 "role" to "user",
-                "content" to "$prompt\n\n$gitData"
+                "content" to prompt
             )
         )
 
         val payload = mapOf(
             "model" to model,
             "messages" to messages,
-            "system" to ROLE,
+            "system" to role,
             "max_tokens" to 500
         )
