+++ src/main/kotlin/com/ibrahimharoon/gitautocommit/llm/service/DefaultLlmResponseService.kt
@@ -1,10 +1,6 @@
 package com.ibrahimharoon.gitautocommit.services
 
-import com.ibrahimharoon.gitautocommit.cache.RegenConversationCache
 import com.ibrahimharoon.gitautocommit.rest.dtos.DefaultLlmResponseDto
-import com.ibrahimharoon.gitautocommit.services.LlmConstants.Companion.COMMIT_PROMPT
-import com.ibrahimharoon.gitautocommit.services.LlmConstants.Companion.PR_SUMMARY_PROMPT
-import com.ibrahimharoon.gitautocommit.services.LlmConstants.Companion.ROLE
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpMethod
@@ -16,26 +12,17 @@ class DefaultLlmResponseService(
     private val url: String,
     private val headers: HttpHeaders
 ) : LlmResponseService {
-    private val conversationCache = RegenConversationCache
-
-    override fun getMessage(gitData: String, isPr: Boolean, additionalLlmPrompt: String): String {
-        var prompt = ""
-        if (additionalLlmPrompt.isNotEmpty()) {
-            prompt = "Your previous commit message was no good. This is an additional prompt to help you out: $additionalLlmPrompt\n\n"
-        }
-        prompt += if (isPr) PR_SUMMARY_PROMPT else COMMIT_PROMPT
-        prompt += if (conversationCache.isNotEmpty()) "Previous history: $conversationCache" else ""
-
+    override fun response(role: String, prompt: String): String {
         val payload = mapOf(
             "model" to model,
             "messages" to listOf(
                 mapOf(
                     "role" to "system",
-                    "content" to ROLE
+                    "content" to role
                 ),
                 mapOf(
                     "role" to "user",
-                    "content" to "$prompt + \n + $gitData"
+                    "content" to prompt
                 )
             )
         )
