--- src/main/kotlin/com/ibrahimharoon/gitautocommit/services/LlmResponseService.kt
@@ -1,5 +1,0 @@
-package com.ibrahimharoon.gitautocommit.services
-
-interface LlmResponseService {
-    fun getMessage(gitData: String, isPr: Boolean, additionalLlmPrompt: String): String
-}
